+++ src/error-boundary.tsx
@@ -1,8 +1,26 @@
 import React, { ErrorInfo } from "react"
+import styled from "styled-components/macro"
+
+const ErrorWrapper = styled.div`
+  width: 100%;
+  height: 100%;
+  font-size: 2rem;
+  padding: 1rem;
+
+  & > h1,
+  p {
+    width: 100%;
+    text-align: center;
+  }
+
+  & > p {
+    font-size: 2rem;
+  }
+`
 
 export class ErrorBoundary extends React.Component {
   state = {
-    hasError: false,
+    hasError: false
   }
 
   static getDerivedStateFromError(error: Error) {
@@ -18,7 +36,18 @@ export class ErrorBoundary extends React.Component {
   render() {
     if (this.state.hasError) {
       // You can render any custom fallback UI
-      return <h1>Something went wrong 🤷‍♂️</h1>
+      return (
+        <ErrorWrapper>
+          <h1>Error</h1>
+          <p>
+            Something went wrong
+            <br />
+            and we don't know why yet 🤷‍♂️
+            <br />
+            <small>(sorry)</small>
+          </p>
+        </ErrorWrapper>
+      )
     }
 
     return <>{this.props.children}</>