+++ src/features/home/components/header/header.styles.ts
@@ -2,13 +2,12 @@ import { Link } from "react-router-dom"
 import { TABLET_MIN_WIDTH } from "shared/constants"
 import styled from "styled-components/macro"
 
-export const Wrapper = styled.header`
+export const StyledWrapper = styled.header`
   width: 100%;
   display: flex;
   white-space: nowrap;
-  column-gap: 3rem;
   align-items: baseline;
-  justify-content: end;
+  justify-content: space-between;
   color: var(--c-gray);
   font-size: 1rem;
   padding: 1rem;
@@ -21,6 +20,18 @@ export const Wrapper = styled.header`
   @media screen and (hover: none) and (orientation: landscape) {
     display: none;
   }
+`
+
+export const StyledDiv = styled.div`
+  display: flex;
+  align-items: baseline;
+  column-gap: 2rem;
+`
+
+export const StyledConnectionDiv = styled.div`
+  display: flex;
+  align-items: center;
+  column-gap: 2rem;
 `
 
 export const StyledLink = styled(Link)`